Bug 1835833: Fix for topology crash on reload
