+++ frontend/packages/dev-console/src/components/topology/filters/index.ts
@@ -1,3 +1,4 @@
+export * from './filter-types';
 export * from './filter-utils';
 export * from './useSearchFilter';
 export * from './useDisplayFilters';
